Adding 3 more stress_cmds for write & read to work with compression (all types).
These cmds use new keyspaces, that's why I need to push the keyspace_name in case compression is in the stress cmd.
There might be issues to work with all possible combination of the test including multi keyspace in round_robin.
We may want to seperate the multi-keyspace logic to a different test case in the future and reduce some of the login in test_custom_time.

(cherry picked from commit 438076a9396627ef63fbf18105b7869d7cae1cb2)
Signed-off-by: Amos Kong <amos@scylladb.com>

Conflicts:
	longevity_test.py
	tests/longevity-50GB-4days.yaml