support to assign google storage url for update_db_packages
We can use gsutil to download directory, we don't care of the package names.
But this requires to install gsutil from google-cloud-sdk repo on db nodes.

Another solution is preparing gsutil in hydra docker, and download the rpms
directory to SCT local first.

gsutil doesn't support to download with Anonymous. We have to config google
cloud account first.

Signed-off-by: Amos Kong <amos@scylladb.com>