cluster_aws: retry to stop scylla-server after instance restart
Currently we stop scylla-server when ssh is up after instance restarts.
It works well most of time, but in some special case, the instance network is
unavailable again in stopping scylla-server, it causes the nemesis fails and
leave a problem instance (no RAID setup).

This patch added a retry strategy for stopping scylla-server after instance
restart.

Signed-off-by: Amos Kong <amos@scylladb.com>